+++ mobile/android/android-components/docs/changelog.md
@@ -19,6 +19,9 @@ permalink: /changelog/
   * Adds `Resources.Theme.resolveAttribute(Int)` to quickly get a resource ID from a theme.
   * Adds `Context.getColorFromAttr` to get a color int from an attribute.
 
+* **feature-customtabs**
+  * Added `CustomTabWindowFeature` to handle windows inside custom tabs, PWAs, and TWAs.
+
 # 14.0.1
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v14.0.0...v14.0.1)